requirements.txt
torch>=2.0.0
numpy>=1.21.0
tokenizers>=0.13.0
transformers>=4.25.0
datasets>=2.8.0
wandb>=0.13.0
tqdm>=4.64.0
pyyaml>=6.0
sentencepiece>=0.1.97
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
tensorboard>=2.10.0
accelerate>=0.15.0
bitsandbytes>=0.37.0
deepspeed>=0.8.0
